+++ projects/examples/source/02.08.container.vector.cpp
@@ -17,7 +17,7 @@ int main()
 
 	for (auto i = size; auto & element : vector)
     {
-        element = static_cast < int > (i--);
+        element = i--;
     }
 
 //  -----------------------------------------------
